radar_service/requirements.txt
asyncpg==0.29.0
nats-py==2.6.0
prometheus-client==0.19.0
numpy==1.24.3
fastapi==0.104.1
uvicorn==0.24.0 